requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
aiosqlite==0.20.0
yoomoney==0.1.2
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9